refactor(S-3.02): extract handle_search to assets/search.rs
Moves handle_search + format_inline_attributes (search-private helper)
to search.rs shard. The --open colorName filter lives in tickets.rs scope;
search.rs owns attribute enrichment. All 612 unit tests pass.
